Rework a bit of the movement code
I've made the movement feel a bit more natural. Before the movement felt
really... floaty? But now he kind of moves and then friction-y comes to a stop.
I have an idea of how to make that code a bit more 'correct' but it looks ok for
now.

I also made it so when not moving jake spins to the front!